requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
transformers==4.35.0
torch==2.3.0
requests==2.31.0
pydantic==2.4.2
cryptography==41.0.7
python-json-logger==2.0.7
pytest==7.4.3
black==23.9.1
flake8==6.1.0
